Separate events generation into separate file for foreman
Alter generate_events.rb to take different starting event_id so multiple
workers can be run off the same script
